requirements.txt
groq
requests
python-dotenv
web3
eth-account
